split unzip to separate file
