+++ pkg/enex/methods.go
@@ -1,9 +1,7 @@
 package enex
 
 import (
-	"archive/zip"
 	"bufio"
-	"bytes"
 	"encoding/base64"
 	"encoding/xml"
 	"enex2paperless/internal/config"
@@ -394,107 +392,6 @@ func (e *EnexFile) RetryFeeder(failedNotes *[]Note) {
 		e.NoteChannel <- note
 	}
 	close(e.NoteChannel)
-}
-
-// isSystemFile checks if a file or directory is a system file that should be excluded
-func isSystemFile(name string) bool {
-	// Convert to lowercase for case-insensitive comparison
-	name = strings.ToLower(name)
-
-	// Check for common system files and directories
-	systemFiles := []string{
-		".ds_store",
-		"thumbs.db",
-		"desktop.ini",
-		"__macosx",
-		"._",
-	}
-
-	for _, systemFile := range systemFiles {
-		if strings.Contains(name, systemFile) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// ExtractedFile represents a file extracted from a zip archive
-type ExtractedFile struct {
-	Path        string
-	Name        string
-	Data        []byte
-	MimeType    string
-	ZipFileName string
-}
-
-// unzipFile takes a byte slice of a zip file and extracts its contents to the specified directory
-// It returns a slice of extracted files with their paths and data
-func unzipFile(data []byte, destDir string, fs afero.Fs, zipFileName string) ([]ExtractedFile, error) {
-	var extractedFiles []ExtractedFile
-
-	// Create a reader from the byte slice
-	zipReader, err := zip.NewReader(bytes.NewReader(data), int64(len(data)))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create zip reader: %v", err)
-	}
-
-	// Create destination directory if it doesn't exist
-	if err := fs.MkdirAll(destDir, 0755); err != nil {
-		return nil, fmt.Errorf("failed to create destination directory: %v", err)
-	}
-
-	// Extract each file
-	for _, file := range zipReader.File {
-		// Skip directories and system files
-		if file.FileInfo().IsDir() || isSystemFile(file.Name) {
-			slog.Debug("skipping system file or directory", "file", file.Name)
-			continue
-		}
-
-		// Open the file in the zip
-		rc, err := file.Open()
-		if err != nil {
-			return extractedFiles, fmt.Errorf("failed to open file in zip: %v", err)
-		}
-
-		// Create the file path
-		filePath := filepath.Join(destDir, file.Name)
-
-		// Read the file contents
-		var buf bytes.Buffer
-		_, err = io.Copy(&buf, rc)
-		rc.Close()
-		if err != nil {
-			return extractedFiles, fmt.Errorf("failed to read file contents: %v", err)
-		}
-
-		// Create the file
-		f, err := fs.Create(filePath)
-		if err != nil {
-			return extractedFiles, fmt.Errorf("failed to create file: %v", err)
-		}
-
-		// Write the contents
-		_, err = f.Write(buf.Bytes())
-		f.Close()
-		if err != nil {
-			return extractedFiles, fmt.Errorf("failed to write file contents: %v", err)
-		}
-
-		// Add file to extracted files list
-		extractedFiles = append(extractedFiles, ExtractedFile{
-			Path:        filePath,
-			Name:        file.Name,
-			Data:        buf.Bytes(),
-			MimeType:    getMimeType(file.Name),
-			ZipFileName: zipFileName,
-		})
-
-		slog.Info("extracted file from zip", "file", file.Name)
-	}
-
-	return extractedFiles, nil
 }
 
 // getMimeType returns the MIME type based on file extension
